+++ skopt/tests/test_searchcv.py
@@ -99,18 +99,20 @@ def test_searchcv_runs_multiple_subspaces():
         'model__C': Real(1e-6, 1e+6, prior='log-uniform'),
     }
 
+    # example subspace with short - hand notation
     dtc_search = {
-        'model': Categorical([DecisionTreeClassifier()]),
-        'model__max_depth': Integer(1, 32),
-        'model__min_samples_split': Real(1e-3, 1.0, prior='log-uniform'),
+        'model': [DecisionTreeClassifier()],
+        'model__max_depth': (1, 32),
+        'model__min_samples_split': (1e-3, 1.0, 'log-uniform'),
     }
 
+    # mixed short - hand and full notations
     svc_search = {
         'model': Categorical([SVC()]),
         'model__C': Real(1e-6, 1e+6, prior='log-uniform'),
-        'model__gamma': Real(1e-6, 1e+1, prior='log-uniform'),
+        'model__gamma': (1e-6, 1e+1, 'log-uniform'),
         'model__degree': Integer(1, 8),
-        'model__kernel': Categorical(['linear', 'poly', 'rbf']),
+        'model__kernel': ['linear', 'poly', 'rbf'],
     }
 
     opt = BayesSearchCV(
@@ -119,8 +121,114 @@ def test_searchcv_runs_multiple_subspaces():
         n_iter=2
     )
 
+    # run the search over subspaces
     opt.fit(X_train, y_train)
 
     # test if all subspaces are explored
     total_evaluations = len(opt.cv_results_['mean_test_score'])
     assert total_evaluations == 1+1+2, "Not all spaces were explored!"
+
+
+def test_searchcv_iterations():
+    """
+    Test whether the `monitor` callbacks are called in BayesSearchCV
+    on space exploration with proper values of total number of
+    iterations to explore and current number of iterations.
+    """
+
+    X, y = load_iris(True)
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, train_size=0.75, random_state=0
+    )
+
+    # used to try different model classes
+    pipe = Pipeline([
+        ('model', SVC())
+    ])
+
+    # single categorical value of 'model' parameter sets the model class
+    lin_search = {
+        'model': Categorical([LinearSVC()]),
+        'model__C': Real(1e-6, 1e+6, prior='log-uniform'),
+    }
+
+    # example subspace with short - hand notation
+    dtc_search = {
+        'model': [DecisionTreeClassifier()],
+        'model__max_depth': (1, 32),
+        'model__min_samples_split': (1e-3, 1.0, 'log-uniform'),
+    }
+
+    # mixed short - hand and full notations
+    svc_search = {
+        'model': Categorical([SVC()]),
+        'model__C': Real(1e-6, 1e+6, prior='log-uniform'),
+        'model__gamma': (1e-6, 1e+1, 'log-uniform'),
+        'model__degree': Integer(1, 8),
+        'model__kernel': ['linear', 'poly', 'rbf'],
+    }
+
+    # a few values that should persist among function calls
+    persistent_values = {
+        'calls_counter': 0
+    }
+
+    def on_step(optim_result):
+        persistent_values['calls_counter'] += 1
+        iter = optim_result.searchcv_iter
+        assert iter == persistent_values['calls_counter']
+
+
+    opt = BayesSearchCV(
+        pipe,
+        [(lin_search, 1), (dtc_search, 1), svc_search],
+        n_iter=2,
+    )
+
+    # run the search over subspaces
+    opt.fit(X_train, y_train, callback=on_step)
+
+    # test if for all iterations function is called
+    total_evaluations = len(opt.cv_results_['mean_test_score'])
+    assert persistent_values['calls_counter'] == total_evaluations
+    assert persistent_values['calls_counter'] == opt.total_iterations
+
+
+def test_searchcv_fit_count():
+    """
+    Test whether BayesSearchCV does not do any unnecessary
+    fitting.
+    """
+
+    X, y = load_iris(True)
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, train_size=0.75, random_state=0
+    )
+
+    # class used to record the number of fittings
+    persistent_values = {'fits_counter':0}
+    n_iter = 2
+
+    class MySVC(LinearSVC):
+        def __init__(self, C=0.1):
+            LinearSVC.__init__(self, C=C)
+
+        def fit(self, X, y, sample_weight=None):
+            persistent_values['fits_counter'] += 1
+            LinearSVC.fit(self, X, y, sample_weight)
+
+    model = BayesSearchCV(
+        estimator=MySVC(),
+        search_spaces={
+            'C': (0.1, 1.0)
+        },
+        n_iter=n_iter,
+        refit=True
+    )
+
+    model.fit(X_train, y_train)
+
+    # 3 cross - validation folds fits for every iteration
+    # + one final fit
+    assert persistent_values['fits_counter']==n_iter*3 + 1
+